replay viewer: add 1/2x playback speed; pin the board page's Space pause
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that spends one sim tick every other frame,
the chrome grows a 0.5x chip, and the frame JSON's sp field carries 0.5 while
it is active (buildStateJson's speed is a float now). '-' floors on 1/2x
instead of 1x, and the live loop's existing clamp keeps a broadcast at 1x.

The lull boost still outranks the parity: a dead stretch is the one place
nobody wants to watch at half pace.

Space already paused on client/replay_broadcast.html, which is the ONLY page
this repo ships -- Dockerfile.replay-viewer builds index.html from it and
there is no league shell here to forward the key down a command channel. The
binding is now pinned by tests/test_viewer.nim rather than re-added.

chrome_common.js keeps reading window.CTF_WIRE: that name is aliased in
wire_constants.nim by design and pinned by test_viewer's "no CTF_ identifier"
sweep, so the shared chrome's only edit is the two-line speed patch (0.5 in
the file:// fallback list, 0.5: '5' in the speed->command map). The byte pin
moves to 594ed4a7 with the reason written into the comment.

Validation: tests/*.nim all green in debug and release (18 files, both modes);
nim c -d:release of src/pistonball.nim and src/pistonball_player.nim; nim
check -d:emscripten replay-viewer/pistonball_replay.nim; gen_wire_constants
emits speeds:[0.5,1,2,3,4,8,16].

+++ src/pistonball/broadcast.nim
@@ -175,7 +175,7 @@ proc buildStateJson*(
   sim: SimServer,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  speed: float,      # the speed the chrome SHOWS: 0.5 at the replay's 1/2x
   maxTick: int,
   looping: bool,
   transportEnabled: bool,
